Check for "alert" param
It is noted when setting the content text that OneSignal uses alert for the message. In my case, "Alert" is the only parameter sent by Parse Dashboard, so the notification is not shown.

I've added a check for the "alert" param along with the other content checks.

The statements were getting a bit unwieldy and hard to read, so I split them into separate lines.